chore: move database from incubation to metadata
